// constants/cryptos.enum.ts
enum ECrypto {
  '00' = '00',
  '1inch' = '1inch',
  aave = 'aave',
  abt = 'abt',
  ach = 'ach',
  acs = 'acs',
  ada = 'ada',
  aergo = 'aergo',
  aero = 'aero',
  agld = 'agld',
  aioz = 'aioz',
  alcx = 'alcx',
  aleph = 'aleph',
  algo = 'algo',
  alice = 'alice',
  akt = 'akt',
  amp = 'amp',
  ankr = 'ankr',
  ant = 'ant',
  ape = 'ape',
  api3 = 'api3',
  apt = 'apt',
  arb = 'arb',
  arkm = 'arkm',
  arpa = 'arpa',
  asm = 'asm',
  ast = 'ast',
  ata = 'ata',
  atom = 'atom',
  auction = 'auction',
  audio = 'audio',
  aurora = 'aurora',
  avax = 'avax',
  avt = 'avt',
  axl = 'axl',
  axs = 'axs',
  badger = 'badger',
  bal = 'bal',
  band = 'band',
  bat = 'bat',
  bch = 'bch',
  bico = 'bico',
  bigtime = 'bigtime',
  bit = 'bit',
  blur = 'blur',
  blz = 'blz',
  bnt = 'bnt',
  boba = 'boba',
  bond = 'bond',
  bonk = 'bonk',
  btc = 'btc',
  btrst = 'btrst',
  busd = 'busd',
  c98 = 'c98',
  cbeth = 'cbeth',
  celr = 'celr',
  cgld = 'cgld',
  chz = 'chz',
  clv = 'clv',
  comp = 'comp',
  corechain = 'corechain',
  coti = 'coti',
  coval = 'coval',
  cro = 'cro',
  crpt = 'crpt',
  crv = 'crv',
  ctsi = 'ctsi',
  ctx = 'ctx',
  cvc = 'cvc',
  cvx = 'cvx',
  dai = 'dai',
  dar = 'dar',
  dash = 'dash',
  ddx = 'ddx',
  deso = 'deso',
  dext = 'dext',
  dia = 'dia',
  dimo = 'dimo',
  dnt = 'dnt',
  doge = 'doge',
  dot = 'dot',
  drep = 'drep',
  dyp = 'dyp',
  egld = 'egld',
  ela = 'ela',
  enj = 'enj',
  ens = 'ens',
  eos = 'eos',
  ern = 'ern',
  etc = 'etc',
  eth = 'eth',
  eurc = 'eurc',
  euroc = 'euroc',
  farm = 'farm',
  fet = 'fet',
  fida = 'fida',
  fil = 'fil',
  fis = 'fis',
  flow = 'flow',
  flr = 'flr',
  fort = 'fort',
  forth = 'forth',
  fox = 'fox',
  fx = 'fx',
  gal = 'gal',
  gala = 'gala',
  gfi = 'gfi',
  ghst = 'ghst',
  glm = 'glm',
  gmt = 'gmt',
  gno = 'gno',
  gnt = 'gnt',
  gods = 'gods',
  grt = 'grt',
  gst = 'gst',
  gtc = 'gtc',
  gusd = 'gusd',
  gyen = 'gyen',
  hbar = 'hbar',
  hft = 'hft',
  high = 'high',
  hnt = 'hnt',
  honey = 'honey',
  hopr = 'hopr',
  icp = 'icp',
  idex = 'idex',
  ilv = 'ilv',
  imx = 'imx',
  index = 'index',
  inj = 'inj',
  inv = 'inv',
  iotx = 'iotx',
  jasmy = 'jasmy',
  jto = 'jto',
  jup = 'jup',
  karrat = 'karrat',
  kava = 'kava',
  keep = 'keep',
  knc = 'knc',
  krl = 'krl',
  ksm = 'ksm',
  lcx = 'lcx',
  ldo = 'ldo',
  link = 'link',
  lit = 'lit',
  loka = 'loka',
  loom = 'loom',
  lpt = 'lpt',
  lqty = 'lqty',
  lrc = 'lrc',
  lseth = 'lseth',
  ltc = 'ltc',
  mana = 'mana',
  mask = 'mask',
  magic = 'magic',
  math = 'math',
  matic = 'matic',
  mco2 = 'mco2',
  mdt = 'mdt',
  media = 'media',
  metis = 'metis',
  mina = 'mina',
  mir = 'mir',
  mkr = 'mkr',
  mln = 'mln',
  mnde = 'mnde',
  mobile = 'mobile',
  mona = 'mona',
  mpl = 'mpl',
  msol = 'msol',
  mtl = 'mtl',
  multi = 'multi',
  musd = 'musd',
  muse = 'muse',
  mxc = 'mxc',
  nct = 'nct',
  near = 'near',
  neon = 'neon',
  nest = 'nest',
  nkn = 'nkn',
  nmr = 'nmr',
  nu = 'nu',
  ocean = 'ocean',
  ogn = 'ogn',
  omg = 'omg',
  ondo = 'ondo',
  ooki = 'ooki',
  op = 'op',
  orca = 'orca',
  orn = 'orn',
  osmo = 'osmo',
  oxt = 'oxt',
  pax = 'pax',
  perp = 'perp',
  pla = 'pla',
  plu = 'plu',
  png = 'png',
  pols = 'pols',
  poly = 'poly',
  pond = 'pond',
  powr = 'powr',
  prime = 'prime',
  pro = 'pro',
  prq = 'prq',
  pundix = 'pundix',
  pyr = 'pyr',
  pyusd = 'pyusd',
  qi = 'qi',
  qnt = 'qnt',
  quick = 'quick',
  rad = 'rad',
  rai = 'rai',
  rare = 'rare',
  rari = 'rari',
  rbn = 'rbn',
  ren = 'ren',
  render = 'render',
  rep = 'rep',
  req = 'req',
  rgt = 'rgt',
  rlc = 'rlc',
  rly = 'rly',
  rndr = 'rndr',
  ronin = 'ronin',
  rose = 'rose',
  rpl = 'rpl',
  safe = 'safe',
  sand = 'sand',
  seam = 'seam',
  sei = 'sei',
  shdw = 'shdw',
  shib = 'shib',
  shping = 'shping',
  skl = 'skl',
  snt = 'snt',
  snx = 'snx',
  sol = 'sol',
  spa = 'spa',
  spell = 'spell',
  stg = 'stg',
  storj = 'storj',
  strk = 'strk',
  stx = 'stx',
  sui = 'sui',
  suku = 'suku',
  super = 'super',
  sushi = 'sushi',
  swftc = 'swftc',
  sylo = 'sylo',
  syn = 'syn',
  t = 't',
  tbtc = 'tbtc',
  tia = 'tia',
  time = 'time',
  tnsr = 'tnsr',
  tone = 'tone',
  trac = 'trac',
  trb = 'trb',
  tribe = 'tribe',
  tru = 'tru',
  tvk = 'tvk',
  uma = 'uma',
  unfi = 'unfi',
  uni = 'uni',
  upi = 'upi',
  usdc = 'usdc',
  usdt = 'usdt',
  ust = 'ust',
  vara = 'vara',
  velo = 'velo',
  vet = 'vet',
  vgx = 'vgx',
  voxel = 'voxel',
  vtho = 'vtho',
  wampl = 'wampl',
  waxl = 'waxl',
  wbtc = 'wbtc',
  wcfg = 'wcfg',
  wluna = 'wluna',
  xcn = 'xcn',
  xlm = 'xlm',
  xmon = 'xmon',
  xrp = 'xrp',
  xtz = 'xtz',
  xyo = 'xyo',
  yfi = 'yfi',
  yfii = 'yfii',
  zec = 'zec',
  zen = 'zen',
  zeta = 'zeta',
  zrx = 'zrx',
};